+++ src/mcmCompiler/include/mcm/tensor/order/order.hpp
@@ -54,8 +54,8 @@ namespace mv
         const std::vector<std::size_t>& getContiguityVector();
 
         // Strides computed in WORDS and bytes respectively
-        std::vector<unsigned> computeWordStrides(const Shape &shape) const;
-        std::vector<unsigned> computeByteStrides(const Shape &s, unsigned dataSize) const;
+        std::vector<float> computeWordStrides(const Shape &shape) const;
+        std::vector<float> computeByteStrides(const Shape &s, unsigned dataSize) const;
         std::size_t operator[](std::size_t idx) const;
         std::size_t size() const;
         std::string toString() const;
